# Conflicts:
#	.idea/compiler.xml
#	.idea/misc.xml
#	.idea/modules.xml

+++ src/ua/gotsman/sms/Sms.java
@@ -15,23 +15,38 @@ import java.nio.file.Paths;
  * Created by Денис on 04/16/2016.
  * Формирует СМС и отправляет его на шлюз smsc.ua через SOAP протокол
  */
-public class Sms {
+public class Sms implements Runnable {
     private final static String LOGIN = "Severik";
     private final static String PASSWORD = "Derparol12!@";
     private final static String SENDER = "SoftTechno";
     private final static String TIME = "0";
-
-    public void sendSms(Info info) throws IOException, BiffException {
+    static int stopTime = 0;
+    @Override
+    public void run() {
+        Info info = new Info();
         Service service = new Service();
         ServiceSoap port = service.getServiceSoap();
         Path path = Paths.get("D:\\1.xls");
         while (true) {
+            if (stopTime == 1) break;
+            try {
+                Thread.sleep(1000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            System.out.println("Test");
             if (Files.exists(path)) {
-                InputStream inputStream = new FileInputStream(String.valueOf(path));
-                info.loadFromXls(inputStream);
-                inputStream.close();
-                Files.delete(path);
+                try {
+                    InputStream inputStream = new FileInputStream(String.valueOf(path));
+                    info.loadFromXls(inputStream);
+                    inputStream.close();
+                    Files.delete(path);
+                } catch (IOException | BiffException e) {
+                    System.out.println(e);
+                }
             }
         }
+        System.out.println("stopped");
+        stopTime = 0;
     }
 }
